# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight, flexible, elegant and full-featured AccountModules scaffolding based on GIN + GORM 2.0 + Casbin 2.0 + Wire DI.",
        "title": "ginadmin",
        "contact": {},
        "version": "v10.0.2"
    },
    "paths": {
...
        "/api/v1/account/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Query role list",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "pagination index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "pagination size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Display name of role",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status of role (disabled, enabled)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.Role"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Create role record",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.RoleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.Role"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/account/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Get role record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.Role"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Update role record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.RoleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "RoleAPI"
                ],
                "summary": "Delete role record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
...
        "/api/v1/p/menus": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Query menu tree data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code path of menu (like xxx.xxx.xxx)",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of menu",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to include menu resources",
                        "name": "includeResources",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.Menu"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "Child menus",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Menu"
                    }
                },
                "code": {
                    "description": "Code of menu (unique for each level)",
                    "type": "string"
                },
                "created_at": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about menu",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string"
                },
                "parent_id": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "parent_path": {
                    "description": "Parent path (split by .)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "properties": {
                    "description": "Properties of menu (JSON)",
                    "type": "string"
                },
                "public": {
                    "description": "Parent path (split by .)",
                    "type": "boolean"
                },
                "resources": {
                    "description": "Resources of menu",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.MenuResource"
                    }
                },
                "sequence": {
                    "description": "Sequence for sorting (Order by desc)",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string"
                },
                "type": {
                    "description": "Type of menu (page, button)",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
...
        "entities.MenuResource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Create time",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menu_id": {
                    "description": "From Menu.ID",
                    "type": "string"
                },
                "method": {
                    "description": "HTTP method",
                    "type": "string"
                },
                "path": {
                    "description": "API request path (e.g. /api/v1/users/:id)",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
...
        "entities.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string"
                },
                "created_at": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menus": {
                    "description": "Role menu list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.RoleMenu"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string"
                },
                "sequence": {
                    "description": "Sequence for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (disabled, enabled)",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
        "entities.RoleForm": {
            "type": "object",
            "required": [
                "code",
                "name",
                "status"
            ],
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string",
                    "maxLength": 32
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "menus": {
                    "description": "Role menu list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.RoleMenu"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string",
                    "maxLength": 128
                },
                "sequence": {
                    "description": "Sequence for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (enabled, disabled)",
                    "type": "string",
                    "enum": [
                        "disabled",
                        "enabled"
                    ]
                }
            }
        },
        "entities.RoleMenu": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Create time",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menu_id": {
                    "description": "From Menu.ID",
                    "type": "string"
                },
                "role_id": {
                    "description": "From Role.ID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
...
        "errors.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "util.ResponseResult": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errors.Error"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}